perf(bundle): lazy-load Defuddle extractor; -90 KB gzip first paint
Values-audit finding #4. The five-year-old phone on slow 4G is the
bottom of the analytics graph. Every byte we ship on first paint is
budget against that user. The audit memo (mistakenly) flagged the
feed catalog as the heavy chunk — that's already lazy-loaded.
The real eager-load weight was Defuddle's full-text extractor.

Defuddle (the HTML pipeline + DOM cleaner) was pulled into the main
bundle via:

  reader-panel.tsx → @/core/extractor/extractor.ts
    → ./adapters/index.ts
      → ./adapters/default-adapter.ts
        → ./defuddle-extractor.ts → import Defuddle from "defuddle"

Reader-panel only needs the small `needsExtraction()` predicate to
decide whether to offer the "Extracted" toggle. The actual extraction
runs when the user explicitly clicks the toggle — i.e. it's a
user-triggered, infrequent action.

Changes:
  * New src/core/extractor/needs-extraction.ts — 20-line predicate,
    no imports from the heavy pipeline.
  * reader-panel.tsx imports the predicate from the new module.
  * extractor.ts re-exports needsExtraction so existing callers
    don't break.
  * extraction-store.ts dynamically imports the extractor + adapter
    registry inside fetchExtracted(). Defuddle now lives in its own
    chunk loaded only when the user invokes extraction.

Bundle impact (production build):
  Before: index.js  1302 KB / 404 KB gzip
  After:  index.js   988 KB / 314 KB gzip   (-90 KB gzip / -22%)

  The 270 KB extractor chunk + 42 KB adapters chunk now load on
  demand. A user who never clicks "Extracted" never pays for them.
  The toggle still feels instant — one network round-trip behind the
  click.

Tests: 2 unit + 3 parity tests updated to await the now-async dispatch
(viewMode flips synchronously; fetch fires after the dynamic import
resolves). Full suite: 2497 passing, 0 type errors.
