+++ src/core/extractor/extractor.ts
@@ -1,8 +1,12 @@
 import { registry } from "./adapters/index.ts";
 import { defaultAdapter } from "./adapters/default-adapter.ts";
-import type { Article } from "../../types/index.ts";
 import type { Result } from "../../utils/result.ts";
 import type { ExtractionResult } from "./defuddle-extractor.ts";
+
+// Re-export so existing call sites keep working. The implementation
+// lives in `./needs-extraction.ts` to keep the Defuddle-free predicate
+// loadable without pulling in the full extractor pipeline.
+export { needsExtraction } from "./needs-extraction.ts";
 
 /**
  * Extract readable content from fetched text.
@@ -12,22 +16,4 @@ import type { ExtractionResult } from "./defuddle-extractor.ts";
 export function extract(text: string, url: string): Result<ExtractionResult> {
   const adapter = registry.findAdapter(url) ?? defaultAdapter;
   return adapter.extract(text, url);
-}
-
-/**
- * Determine whether an article needs full-text extraction.
- * Returns true if the article appears to only have a summary/teaser.
- */
-export function needsExtraction(article: Article): boolean {
-  if (!article.link || !article.link.startsWith("http")) return false;
-
-  const content = article.content || "";
-  const summary = article.summary || "";
-
-  // Has distinct, non-trivial content — no extraction needed
-  if (content && content !== summary) return false;
-
-  // Content is empty or identical to summary — check if summary is short enough
-  // to be a teaser rather than a complete short article
-  return summary.length < 500;
 }
